New function Calculate_Mag_Heading()
dRehmFlight already had mag calibration and magx magy magz calculation, but was lacking heading calculation. 
So i am happy to share my implementation of compensated heading calculation, Theoreticly my code should be correct, so far no succes for me.
The code is not yet obtimised and there need to bee some testing before add to the final code but it should work. The code calulates the compensated magnetic north and then with added declination the geografical north. 

Please check the chagedcode.ino file and leave behind your opinion about which of the 2 options is the best.